Change `ServerContributions` type from `decimal` to `float` in `AnnualBudget`.

+++ src/Models.Fog/Entities/AnnualBudget.cs
@@ -3,6 +3,6 @@ namespace Ingweland.Fog.Models.Fog.Entities;
 public class AnnualBudget
 {
     public int Year { get; set; }
-    public decimal ServerContributions { get; set; }
+    public float ServerContributions { get; set; }
     public int ServerGoal { get; set; }
 }
